refactor(Quota): move to typescript
